[Task 1863798] extract abstract entity manager into common lib
